+++ CHANGELOG.md
@@ -32,9 +32,10 @@ to the project during `#hactoberfest`. List of awesome people:
 - Forbid `for` loops with unused `else`
 - Forbid variables self reassignment
 - Forbid `try` with `finally` without `except`
-- Forbit `if` statements with invalid conditionals
+- Forbid `if` statements with invalid conditionals
 - Forbid opening parenthesis from following keyword without space in between them
 - Forbid the use of more than 2 `for` loops within a comprehension
+- Forbid variable names with more than one consecutive underscore
 - Restrict the maximum number of base classes aka mixins
 - Forbid importing protected names